requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Flask-SQLAlchemy==3.0.5
validate-email-address==1.0
python-dotenv==1.0.0
